pyproject.toml
[tool.poetry]
name = "frames-api-grinatom-task"
version = "0.1.0"
description = ""
authors = ["Orkhan Kazymov <rayskarken@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.37"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.0"
uvicorn = "^0.17.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
environs = "^9.5.0"
passlib = "^1.7.4"
python-multipart = "^0.0.5"
itsdangerous = "^2.1.2"
bcrypt = "^3.2.2"
cryptography = "^37.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
